+++ internal/client/transport/kcp.go
@@ -213,7 +213,7 @@ func (c *KcpTransport) Restart() {
 	c.logger.Info("restarting client...")
 
 	// for removing timeout logs
-	level := c.logger.Level
+	level := c.logger.GetLevel()
 	c.logger.SetLevel(logrus.FatalLevel)
 
 	if c.state.Cancel() != nil {