+++ internal/client/transport/quic.go
@@ -129,7 +129,7 @@ func (c *QuicTransport) Restart() {
 	c.logger.Info("restarting client...")
 
 	// for removing timeout logs
-	level := c.logger.Level
+	level := c.logger.GetLevel()
 	c.logger.SetLevel(logrus.FatalLevel)
 
 	if c.state.Cancel() != nil {