+++ internal/client/transport/tcp.go
@@ -118,7 +118,7 @@ func (c *TcpTransport) Restart() {
 	c.logger.Info("restarting client...")
 
 	// for removing timeout logs
-	level := c.logger.Level
+	level := c.logger.GetLevel()
 	c.logger.SetLevel(logrus.FatalLevel)
 
 	if c.state.Cancel() != nil {