+++ internal/client/transport/udp.go
@@ -88,7 +88,7 @@ func (c *UdpTransport) Restart() {
 	c.logger.Info("restarting client...")
 
 	// for removing timeout logs
-	level := c.logger.Level
+	level := c.logger.GetLevel()
 	c.logger.SetLevel(logrus.FatalLevel)
 
 	if c.state.Cancel() != nil {