+++ internal/server/transport/kcp.go
@@ -287,7 +287,7 @@ func (s *KcpTransport) Restart() {
 	s.run.stop()
 
 	// for removing timeout logs
-	level := s.logger.Level
+	level := s.logger.GetLevel()
 	s.logger.SetLevel(logrus.FatalLevel)
 
 	if s.controlChannel.IsSet() {