+++ internal/server/transport/quic.go
@@ -187,7 +187,7 @@ func (s *QuicTransport) Restart() {
 	s.run.stop()
 
 	// for removing timeout logs
-	level := s.logger.Level
+	level := s.logger.GetLevel()
 	s.logger.SetLevel(logrus.FatalLevel)
 
 	if s.controlChannel.IsSet() {