+++ internal/server/transport/tcp.go
@@ -170,7 +170,7 @@ func (s *TcpTransport) Restart() {
 	s.logger.Info("restarting server...")
 
 	// for removing timeout logs
-	level := s.logger.Level
+	level := s.logger.GetLevel()
 	s.logger.SetLevel(logrus.FatalLevel)
 
 	s.run.stop()