+++ internal/server/transport/wsmux.go
@@ -167,7 +167,7 @@ func (s *WsMuxTransport) Restart() {
 	s.logger.Info("restarting server...")
 
 	// for removing timeout logs
-	level := s.logger.Level
+	level := s.logger.GetLevel()
 	s.logger.SetLevel(logrus.FatalLevel)
 
 	s.run.stop()